Preserve attributes set by attr_list
In order to preserve attributes of the image set by the built-in
attr_list extension, yafg has to have a different priority. 8 seems to
be the sweet spot, 9 doesn't work. Note that this is just a hacky
heuristic and may break other plugins.

Additionally, another processor seems to run too early or late now,
because we have to manually add a newline after each figure tag in
order to preserve the current behaviour.

This will haunt me one day.